get_magic_quotes_runtime() is deprecated as of PHP 7.4
Remove call to get_magic_quotes_runtime() as this function is deprecated
as of PHP 7.4 (see https://wiki.php.net/rfc/deprecations_php_7_4). This
part of the code was only needed for PHP <= 5.3 and this version is not
supported anymore by WooCommerce.

Cherry-pick of https://github.com/woocommerce/woocommerce/commit/1f69fbcf3 to Classic Commerce.