Teach Travis to verify freshness of generated (built) files (#269)
* Teach Travis to verify freshness of generated (built) files

* Update admin-rtl.css file

* Add debug info

* More debug info

* Replace grunt-contrib-concat with a custom task

This ensures that the CSS files containing select2.css are only updated
if they need to be, based on their contents. Whatever
grunt-contrib-concat was doing instead didn't always work correctly.

* Update admin CSS

* Fix CSS charset

Per https://www.w3schools.com/cssref/pr_charset_rule.asp "The @charset
rule must be the first element in the style sheet and not be preceded by
any character." This was not previously the case.

* Remove debug code

This makes for very long Travis output, which is harder to read.